Update message_onoff.py : Migrated Message Functions
