+++ message.py
@@ -7,13 +7,11 @@ from message_util.message_logistics import message_logistics
 from message_util.message_meal import message_meal
 from message_util.message_meme import message_meme
 from message_util.message_memory import message_memory
+from message_util.message_onoff import message_onoff
 from util.cipher_util import DESAdapter
 
 import certifi
 import datetime
-import json
-import os
-import random
 import requests
 
 
@@ -56,6 +54,10 @@ def getReplyMessage(message, room, sender):
     if result_message is not None:
         return result_message
 
+    result_message = message_onoff(message, room, sender)
+    if result_message is not None:
+        return result_message
+
     if "학사일정" in message:
         strResult = messageCAUCalendar()
     elif "열람실" in message:
@@ -69,11 +71,6 @@ def getReplyMessage(message, room, sender):
             strResult = messageNSULibrary()
         else:
             strResult = messageCAULibrary("")
-    elif "와봇" in message:
-        if "꺼" in message or "끄" in message:
-            strResult = messageWabotPower(0, room)
-        elif "켜" in message or "키" in message:
-            strResult = messageWabotPower(1, room)
     elif "비트코인" in message:
         strResult = messageBitcoin()
     return strResult
@@ -164,40 +161,5 @@ def messageNSULibrary():
     third = "제3 자유열람실 : 여석 %s석 (%s석 사용중)" % (str(324 - int(Response['data']['data'][2]['inUse']) - int(Response['data']['data'][2]['fix']) - int(Response['data']['data'][2]['disabled'])), Response['data']['data'][2]['inUse'])
 
     strMessage = "남서울대학교 열람실 좌석현황(성암기념중앙도서관)\n\n" + first + second + third
-
-    return strMessage
-
-def messageWabotPower(flag, room):
-    strMessage = ""
-    if os.path.isfile("power.json"):
-        with open('power.json', 'r', encoding='utf-8') as f:
-            power_dict = json.load(f)
-        room_power = power_dict.get(room)
-        if room_power is not None:
-            if (flag == 0 and room_power == "0") or (flag == 1 and room_power == "1"):
-                return strMessage
-        else:
-            if flag == 1: return strMessage
-            else: pass
-    else:
-        power_dict = {}
-    randInt = random.randrange(0, 4)
-    if randInt == 0:
-        if flag == 0:
-            power_dict[room] = "0"
-            json_data = json.dumps(power_dict, ensure_ascii=False, indent=4)
-            strMessage = "와봇이 종료되었습니다."
-        elif flag == 1:
-            power_dict[room] = "1"
-            json_data = json.dumps(power_dict, ensure_ascii=False, indent=4)
-            strMessage = "와봇이 시작되었습니다."
-        with open('power.json', 'w', encoding='utf-8') as f:
-            f.write(json_data)
-    elif randInt == 1:
-        strMessage = "싫은데? ^^"
-    elif randInt == 2:
-        strMessage = "네~"
-    elif randInt == 3:
-        strMessage = "ㅋㅋ"
 
     return strMessage
